# Server/Startup/openapi.json
{
  "x-generator": "NSwag v14.3.0.0 (NJsonSchema v11.2.0.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/api/Ai/chat": {
      "post": {
        "tags": [
          "Ai"
        ],
        "operationId": "Ai_Subscribe",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "message",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageFromClient"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ai/historical_data_analysis": {
      "post": {
        "tags": [
          "Ai"
        ],
        "operationId": "Ai_HistoricalDataAnalysis",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "message",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageFromClient"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Subscription/Subscribe": {
      "post": {
        "tags": [
          "Subscription"
        ],
        "operationId": "Subscription_Subscribe",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeSubscriptionDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Subscription/Unsubscribe": {
      "post": {
        "tags": [
          "Subscription"
        ],
        "operationId": "Subscription_Unsubscribe",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeSubscriptionDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MessageFromClient": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ChangeSubscriptionDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "clientId": {
            "type": "string"
          },
          "topicIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "WebsocketMessage_1": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "topic": {
                "type": "string"
              },
              "data": {
                "$ref": "#/components/schemas/T"
              }
            }
          }
        ]
      },
      "T": {
        "type": "object",
        "additionalProperties": false
      },
      "BaseDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "eventType": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "ApplicationBaseDto": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "eventType": {
            "type": "string"
          },
          "topic": {
            "type": "string"
          }
        }
      },
      "LiveAiFeedbackDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "aiAdvice": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ServerSendsErrorMessage": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ClientRequestAiLiveData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "RequestAirQualityData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "timePeriod": {
                "$ref": "#/components/schemas/TimePeriod"
              },
              "data": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "TimePeriod": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Hourly",
          "Daily",
          "Weekly",
          "Monthly"
        ],
        "enum": [
          "Hourly",
          "Daily",
          "Weekly",
          "Monthly"
        ]
      },
      "ServerResponseAirQualityData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "requestedData": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "timePeriod": {
                "$ref": "#/components/schemas/TimePeriod"
              },
              "data": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        ]
      },
      "ClientRequestDeviceHistory": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "deviceId": {
                "type": "string"
              },
              "pageNumber": {
                "type": "integer",
                "format": "int32"
              },
              "pageSize": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
      "ServerResponseDeviceHistory": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "connectionData": {
                "$ref": "#/components/schemas/PagedResultOfDeviceConnectionHistoryDto"
              }
            }
          }
        ]
      },
      "PagedResultOfDeviceConnectionHistoryDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceConnectionHistoryDto"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "DeviceConnectionHistoryDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "deviceId": {
            "type": "string"
          },
          "isConnected": {
            "type": "boolean"
          },
          "lastSeen": {
            "type": "integer",
            "format": "int64"
          },
          "duration": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        }
      },
      "ClientRequestDeviceStatus": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "ServerResponseDeviceStatus": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "deviceStatus": {
                "$ref": "#/components/schemas/DeviceDto"
              }
            }
          }
        ]
      },
      "DeviceDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "device_id": {
            "type": "string"
          },
          "DeviceName": {
            "type": "string"
          },
          "LastSeen": {
            "type": "integer",
            "format": "int64"
          },
          "IsConnected": {
            "type": "boolean"
          }
        }
      },
      "ClientRequestDeviceStats": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "ServerResponseDeviceStats": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "stats": {
                "$ref": "#/components/schemas/DeviceStatsDto"
              }
            }
          }
        ]
      },
      "DeviceStatsDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "allTimeMeasurements": {
            "type": "integer",
            "format": "int32"
          },
          "connectedDevices": {
            "type": "integer",
            "format": "int32"
          },
          "disconnectionsLast24Hours": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ClientRequestSensorData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "sensorId": {
                "type": "string"
              },
              "pageNumber": {
                "type": "integer",
                "format": "int32"
              },
              "pageSize": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
      "ServerResponseSensorData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "sensorData": {
                "$ref": "#/components/schemas/PagedResultOfSensorDataDto"
              }
            }
          }
        ]
      },
      "PagedResultOfSensorDataDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SensorDataDto"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SensorDataDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "temperature": {
            "type": "number",
            "format": "double"
          },
          "humidity": {
            "type": "number",
            "format": "double"
          },
          "air_quality": {
            "type": "number",
            "format": "double"
          },
          "pm25": {
            "type": "number",
            "format": "double"
          },
          "device_id": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ClientRequestDeviceList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "ServerResponseDeviceList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "deviceList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              }
            }
          }
        ]
      },
      "Ping": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "Pong": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "WebsocketTopics": {
        "type": "string",
        "description": "Websocket Topic Enums",
        "enum": [
          "Dashboard",
          "Ai",
          "DeviceStatus",
          "DeviceData",
          "Device",
          "Graph",
          "GraphTotalMeasurements"
        ]
      },
      "WebsocketEvents": {
        "type": "string",
        "description": "Websocket event types (constants + BaseDto inheritors)",
        "enum": [
          "AllDeviceStatus",
          "BroadcastDeviceConnectionUpdate",
          "BroadcastDeviceSensorDataUpdate",
          "ClientRequestAiLiveData",
          "ClientRequestDeviceHistory",
          "ClientRequestDeviceList",
          "ClientRequestDeviceStats",
          "ClientRequestDeviceStatus",
          "ClientRequestSensorData",
          "DeviceUpdateStatus",
          "GraphGetMeasurement",
          "GraphTemperatureUpdate",
          "GraphTotalMeasurement",
          "LiveAiFeedbackDto",
          "Ping",
          "Pong",
          "RequestAirQualityData",
          "ServerResponseAirQualityData",
          "ServerResponseDeviceHistory",
          "ServerResponseDeviceList",
          "ServerResponseDeviceStats",
          "ServerResponseDeviceStatus",
          "ServerResponseSensorData",
          "ServerSendsErrorMessage",
          "WebsocketMessage`1"
        ]
      }
    }
  }
}